# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ELFOR LED Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ===== ARTICLES =====
...
components:
  schemas:
...
    PriceTier:
      type: object
      required: [minQty, price]
      properties:
        minQty:
          type: integer
        price:
          type: number
    ProductSpec:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string
        unit:
          type: ["string", "null"]
    Product:
      type: object
      required: [id, name, sku, categoryId, price, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        series:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        power:
          type: ["string", "null"]
        lumens:
          type: ["integer", "null"]
        colorTemp:
          type: ["string", "null"]
        ipRating:
          type: ["string", "null"]
        warranty:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        inStock:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, sku, categoryId, price, stock]
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        series:
          type: string
        categoryId:
          type: integer
        shortDescription:
          type: string
        fullDescription:
          type: string
        price:
          type: number
        oldPrice:
          type: number
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        power:
          type: string
        lumens:
          type: integer
        colorTemp:
          type: string
        ipRating:
          type: string
        warranty:
          type: string
        stock:
          type: integer
        featured:
          type: boolean
    ProductPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        series:
          type: ["string", "null"]
        categoryId:
          type: integer
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        priceTiers:
          type: array
          items:
            $ref: "#/components/schemas/PriceTier"
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        power:
          type: ["string", "null"]
        lumens:
          type: ["integer", "null"]
        colorTemp:
          type: ["string", "null"]
        ipRating:
          type: ["string", "null"]
        warranty:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        inStock:
          type: boolean

    # ===== ARTICLE =====
...