# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
import type { PriceTier } from './priceTier';
import type { ProductSpec } from './productSpec';

export interface Product {
  id: number;
  name: string;
  sku: string;
  /** @nullable */
  series?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  fullDescription?: string | null;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  priceTiers?: PriceTier[];
  specs?: ProductSpec[];
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  /** @nullable */
  power?: string | null;
  /** @nullable */
  lumens?: number | null;
  /** @nullable */
  colorTemp?: string | null;
  /** @nullable */
  ipRating?: string | null;
  /** @nullable */
  warranty?: string | null;
  stock: number;
  featured?: boolean;
  inStock?: boolean;
  createdAt: string;
}