# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ELFOR LED Store API
 * OpenAPI spec version: 0.1.0
 */
import type { PriceTier } from './priceTier';
import type { ProductSpec } from './productSpec';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sku: string;
  series?: string;
  categoryId: number;
  shortDescription?: string;
  fullDescription?: string;
  price: number;
  oldPrice?: number;
  priceTiers?: PriceTier[];
  specs?: ProductSpec[];
  imageUrl?: string;
  images?: string[];
  power?: string;
  lumens?: number;
  colorTemp?: string;
  ipRating?: string;
  warranty?: string;
  stock: number;
  featured?: boolean;
}